Moves Node and Edge classes into graph-internals.ts to break cyclic dependencies
